database test

+++ databaseSeed.sql
@@ -12,7 +12,7 @@ CREATE TABLE student
 , address_city          VARCHAR(20)   NOT NULL
 , address_state         VARCHAR(2)    NOT NULL
 , address_zip           INTEGER       NOT NULL
-, grade                 INTEGER(2)    NOT NULL
+, grade                 INTEGER       NOT NULL
 , homeroom_teacher      VARCHAR(20)   NOT NULL);
